requirements.txt
bcrypt==3.2.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
future==0.18.2
google-api-core==2.7.1
google-api-python-client==2.43.0
google-auth==2.6.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.0
httplib2==0.20.4
idna==3.3
netmiko==4.0.0
ntc-templates==3.0.0
oauth2client==4.1.3
paramiko==2.10.3
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDrive==1.3.1
PyNaCl==1.5.0
pyparsing==3.0.8
pyserial==3.5
python-dotenv==0.20.0
PyYAML==6.0
requests==2.27.1
rsa==4.8
schedule==1.1.0
scp==0.14.4
six==1.16.0
tenacity==8.0.1
textfsm==1.1.2
uritemplate==4.1.1
urllib3==1.26.9
